requirements.txt
mysql-connector-python>=8.0.0
pandas>=1.3.0
boto3>=1.26.0 